+++ game/core/src/main/java/com/spacecomplexity/longboilife/game/GameScreen.java
@@ -12,6 +12,7 @@ import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
 import com.spacecomplexity.longboilife.Main;
 import com.spacecomplexity.longboilife.MainInputManager;
+import com.spacecomplexity.longboilife.game.achievements.AchievementsManager;
 import com.spacecomplexity.longboilife.game.building.Building;
 import com.spacecomplexity.longboilife.game.building.BuildingType;
 import com.spacecomplexity.longboilife.game.globals.*;
@@ -34,6 +35,7 @@ public class GameScreen implements Screen {
     private ShapeRenderer shapeRenderer;
     private UIManager ui;
     private InputManager inputManager;
+    private AchievementsManager achievementsManager;
 
     private Viewport viewport;
 
@@ -69,6 +71,9 @@ public class GameScreen implements Screen {
         MainTimer.getTimerManager().getTimer().setEvent(() -> {
             EventHandler.getEventHandler().callEvent(EventHandler.Event.GAME_END);
         });
+
+        // ASSESSMENT 2 - Initialise the achievements manager
+        achievementsManager = new AchievementsManager();
 
         // Create an input multiplexer to handle input from all sources
         InputMultiplexer inputMultiplexer = new InputMultiplexer(new MainInputManager());
@@ -302,6 +307,11 @@ public class GameScreen implements Screen {
         if (!gameState.paused && !MainTimer.getTimerManager().getTimer().poll()) {
             // Update the satisfaction score
             GameUtils.updateSatisfactionScore(world);
+            // ASSESSMENT 2 - Check for achievements and display them
+            achievementsManager.checkAchievements();
+            ui.showAchievement(achievementsManager.getAchievementQueue());
+            // Update the unlocked achievements in the game state
+            gameState.unlockedAchievements = achievementsManager.getUnlockedAchievements();
         }
     }